pyproject.toml
[tool.poetry]
name = "prefect-grupy"
version = "0.1.0"
description = ""
authors = ["Gabriel Montañola <gabriel@montanola.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
prefect = {extras = ["aws", "templates"], version = "^0.14.19"}
pandas = "^1.2.4"
requests = "^2.25.1"
"aws-cdk.core" = "^1.105.0"
"aws-cdk.aws-ec2" = "^1.105.0"
"aws-cdk.aws-ecs" = "^1.105.0"
"aws-cdk.aws-s3" = "^1.105.0"
dask = "^2021.5.0"
distributed = "^2021.5.0"

[tool.poetry.dev-dependencies]
black = "^21.5b1"
isort = "^5.8.0"
flake8 = "^3.9.2"
ipdb = "^0.13.7"
pre-commit = "^2.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
linelength = 100

[tool.isort]
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
